fix: stop measure responsetime for paths without router.
To avoid OOM than can be triggered by measuring a new
label per path, even if it is not handled by the application
is not a smart thing, as it can explode the number of
time dimensions.

Added a check that the request was handled by a defined route
in express, if not the path is set to "(hidden)".