feat: rename :global: env to "default"
Our testing and internal validation has proven that
the :global: environment concept confuses people more
than the problems it solves. We have thus decided to
group all configuration that was created before the
environment concept was introdcued in to the "default
environment. This would still make everything work
as before in addition to introducing the env concept.