create new test:unit script for unit tests
the idea here was to create a new node script target
which ran only the in-memory unit tests which had no
dependency on an external postgres database.

curiously, though, I had to keep adding exclusions
until the tests pass.

in this commit although the test:unit script passes:

    Snapshot Summary
     › 1 snapshot written from 1 test suite.

    Test Suites: 86 passed, 86 total
    Tests:       440 passed, 440 total
    Snapshots:   1 written, 1 passed, 2 total
    Time:        30.022 s
    Ran all test suites.

the console output suggests that a handful of tests
are throwing errors and yet still passing

being new to this codebase I do not yet understand
whether this is expected behaviour or a sign of
a false positive (where the test should be failing
but isn't properly recognizing the error)