docs: Add section on (potentially) stale toggles to tech debt
Previously, when you searched for 'stale toggle' in the docs, the
first result would be the API docs and it'd be hard to find any form
of explanation of what a stale toggle is at all.

To combat that, this commit adds a section to the tech debt chapter
explaining what they are.