docs(fix): fix quickstart image not being displayed.
The issue was that the image link was encased in an extra set of
brackets. However, the extra bracket set was added because docusaurus
wouldn't compile otherwise. Instead it complained about an
unterminated string literal in the alt text.

As it turns out, this had nothing to do with the contents of the alt
text, but rather with the fact that the alt text was split over
multiple lines.