docs: reorder sdks listing in sidebar
The previous list order didn't seem to have any reason behind it. This
is a suggestion for new order.

Index (the introduction) is first, followed by the proxy. Then comes
all the official sdks in alphabetic order (punctuation placed before
alpha). At the end of the list is the 'community sdks' link.

I'm open to suggestions and discussion here, but was working on this
list, so thought it could use some love.