docs(a11y): remove role 'alert' from availability notice.
According to [MDN's article on the `alert`
role](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/Alert_Role),
`alert` should only be used for "information that requires the user's
immediate attention", such as the user's login session being about to
expire or the connection to the server being lost so that local
changes will not be saved.

They also add that:

> Warning: Because of its intrusive nature, the alert role must be used sparingly and only in situations where the user's immediate attention is required.

While availability may be an important (or at least relevant) piece of
information, I do not think it classifies as urgent or requiring
immediate attention.