docs(a11y): dark mode fix contrast in sidebar and with vid comp
This commit introduces a new primary color gradient palette to replace the
dark / mossy green that we use in the light theme. It's got a very
small range, so we should probably expand on it in the future, but for
now it'll do a _lot_ better than it did previously.