docs: adjust primary color scale, fix search dropdown a11y issues
This changes the darker end of the primary color scale to make it even
darker. This means that at the lower end of the scale, white text also
gets enough contrast to be accessible.

For the record: these variables do not seem to be in use anywhere else
in the code base (based on a text search), so it shouldn't affect any
other bits or pieces.