fix: make sure customer type is included in the payload.
Because it previously set it via a reducer and submitted the form in
the same step, the customer type wouldn't be set correctly before the
form was submitted, causing it to show up as "undefined".

We're doing double the work now, but I think that's an acceptable
trade-off for now.