Fix(#1391): fix edge case URLs in code samples
## What

This change fixes URLs in code samples for endpoints that don't start
with /api. Most endpoints do, but there are some edge cases, such as
the edge and auth endpoints that don't. This change makes sure that
those code samples are also correct.

## Why

So that code samples don't present the wrong URL. That'd be really
confusing to users.