docs: Remove/update references to Heroku (#2099)
## What

This PR removes or updates references in the docs to Heroku. Most of the code samples have been replaced with a more generic `unleash.example.com` url, while other references have been removed or updated.

Also removes old OpenAPI files that are out of date and redundant with the new generation.

## Background

Come November and Heroku will no longer offer free deployments of Unleash, so it's about time we remove that claim.

Links to the heroku instance are also outdated because we don't have that instance running anymore.

Finally, the OpenAPI files we do have there are old and static, so they don't match the current reality.

## Commits

* Meta: update ignore file to ignore autogenerated docs

I must've missed the ignore file when looking for patterns.

* docs: delete old openapi file.

This seems to have been a holdover from 2020 and is probably
hand-written. It has been superseded by the new autogenerated OpenAPI docs.

* docs: add notes for heroku changes to the frontend readme and pkg

* docs: remove old openapi article and add redirects to new openapi

* docs: fix link in overview doc: point to GitHub instead of heroku

* docs: update quickstart docs with new heroku details

* docs: remove reference to crashing heroku instance

* docs: remove references to herokuapp in  code samples

* docs: add a placeholder comment

* docs: update references for heroku updates

* docs: keep using unleash4 for enterprise

* docs: remove start:heroku script in favor of start:sandbox

* docs: remove 'deploy on heroku button'

Now that it's not free anymore (or won't be very shortly), let's
remove it.

* docs: remove extra newline