docs: update export/import examples
## What

This change updates the availability notice and the code samples in the import/export docs.

Specifically, the code samples have been moved to using the ApiRequest
component (and thus get examples in HTTP, cURL, and HTTPie), and the
import code sample now says which version it's using (version 3).

## Why

The import code sample didn't say which version of the data it used,
so Unleash would assume it was version 1 if you tried to run it.
However, the actual data was incompatible with the v1 data format, so
you'd get an error if you tried to actually import it.

While fixing this, I figured it'd also be alright to add some other
minor updates to bring the docs a bit more up to recent standards.