chore: update docusaurus/openapi integration to stable version
## What

This change updates the version of the docusaurus-openapi plugin we depend on for openapi docs generation from canary version 0.0.0-514 to stable release 1.4.4.

## Why

When we set the canary version before the weekend it was because it contained a fix of some of the issues we were having with our docs.

These have now been released in v1.4.4, so we can upgrade to a stable version.