docs: various fixes (#2669)
## What

This PR contains a number of minor fixes to the docs in terms of
formatting and redirects. The changes are:

- Adding an availability notice for the front-end API reference docs
(with links to the release blog post).

- Fix malformed admonition headers in the Unleash proxy docs and
impression data docs.

-   Add missing redirects:

- `/advanced/custom_activation_strategy` ->
`/reference/custom-activation-strategies`
- `/docs/deploy/configuring_unleash` ->
`/reference/deploy/configuring-unleash`
- `/docs/user_guide/connect_sdk` -> `/reference/sdks`(this one goes to
the SDKs doc instead of the quickstart because there's more
comprehensive information there and because `/user_guide/connect_sdk`
already goes there)
    -   `/sdks/proxy-javascript` -> `/reference/sdks/javascript-browser`
    -   `/sdks/proxy-react`-> `/reference/sdks/react`
- `/docs/getting_started` -> `/tutorials/quickstart` (this does not go
to "Deploy: getting started" because that is its own link)

## Why

Because keeping the docs up to date and accessible is important.