chore: remove constraints announcements from readme and docs
## What

This change removes the Unleash 4.16 constraints announcement from
readme and from the docs.

## Why

It feels like 4.16 has been out for long enough now. Gearing up for
version 4.19, it may be time to take this away.

## Discussion

However, open-sourcing constraints was (and still is) a big deal, so
it might be worth keeping the banner around, but maybe in a more
subdued format? Something like "did you know Unleash's constraints
feature is open source, now?".