fix: Preserve all data from strategy import
Previously the strategy was to use the IMinimalStrategyRow.
This PR changes to use the full data from the export in the import.

* Discussion point:
Should we see if we can add some invariants to our import/export so that
we can write some proptests for it. One candidate is commutativity of
import/export. On a fresh database importing and then exporting should
yield the same file that was imported provided all flags are turned on